1/5/21
added method setPowerAllMotor() to make the robot move (assumed) and used the variables from the goToPosition() --> i.e. x component, y component, pivot correction.

we have to test once we get the encoders if this code works bc we are basing a lot off of assumptions
ex: wheels just move wiht SetPower() methods

Things for next time:
Rishika and I figured out that this only works if we are in a certain starting position (red far right starting) so we need to make one for red far left starting position (based on the assumption that we are only participating in remote events and can choose (i guess) where we want to start)

cute idea: scavenger hunt in commit messages to teach how them how to use github !

Signed-off-by: Angela Ngo <ngo.angela@gmail.com>